Refactor response structure by removing ID from TableResponse; update GetTables and GetTable functions accordingly. Enhance type inference in InferTypeFromValue to distinguish between int and float types; adjust tests to reflect changes.

+++ internal/utils/eav_test.go
@@ -88,7 +88,7 @@ func TestInferTypeFromValue_Success(t *testing.T) {
 		expected string
 	}{
 		{"hello", "string"},
-		{123, "float"},
+		{123, "int"},
 		{3.14, "float"},
 		{true, "bool"},
 	}
